refactor: 좌표값 VO 분리, 패키지 이동
