+++ src/main/kotlin/minesweeper/domain/cell/Coordinate.kt
@@ -1,0 +1,6 @@
+package minesweeper.domain.cell
+
+data class Coordinate(
+    val x: CoordinateValue,
+    val y: CoordinateValue,
+)
